Big Update
Added menu with instructions and credits (be sure to put your name in
the credits). Added another level and made it so the game map creates a
grid to make it easier to navigate the map. Added a restart function (R
key) so
the player will start from its start position when the map was
loaded.